Refactor KAP20 path Fix typo function and more in KAP721

+++ contracts/security/interface/IKYCBitkubChain.sol
@@ -2,5 +2,5 @@
 pragma solidity ^0.8.0;
 
 interface IKYCBitkubChain {
-    function kycsLevel(address contractAddress) external view returns (uint256);
+    function kycsLevel(address contractAddress) external view returns (uint8);
 }
